+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVVTHHQW037XE1D7HAZFXR1J"
+        "01KVWWW4EPSTZZ8WR5YQSDB2AS"
     )
     r"""The unique identifier of the data source"""