+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVVTHHQQC2W0R4X4FA5WKTW1"
+        "01KVWWW4EHYE7WG0FYT233NN4M"
     )
     r"""The unique identifier of the data source"""