+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KVVTHHK2CY3BK115S4JK5GFV"
+    id: Optional[str] = "01KVWWW4A3WTYXCY297BJ5Q5RB"
 
     description: Optional[str] = None
 
@@ -1357,7 +1357,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KVVTHHJRZ84MK49QDBKQXQN5"
+        "tool_01KVWWW49T3XD4QF1F7TBMJ8J9"
     )
 
     display_name: Optional[str] = None
@@ -1488,7 +1488,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KVVTHHJPR0YDADCK5TVE74BD"
+    id: Optional[str] = "01KVWWW49SNBTE8BNMZAEPCG5G"
 
     description: Optional[str] = None
 
@@ -1632,7 +1632,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KVVTHHJMKNMDDTABR4A7GEFT"
+        "tool_01KVWWW49QC08MHWKMJZHGXG4Y"
     )
 
     display_name: Optional[str] = None
@@ -1932,7 +1932,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KVVTHHJG37GP994YC0ZSTKHP"
+        "tool_01KVWWW49M21XBGN8CVQPXA904"
     )
 
     display_name: Optional[str] = None
@@ -2126,7 +2126,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KVVTHHJEQ12RQYBRBAW87RCC"
+        "tool_01KVWWW49JRW7QXF31F0ZVWH1E"
     )
 
     display_name: Optional[str] = None
@@ -2324,7 +2324,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KVVTHHJBB52G2ANFPFDKHJ52"
+        "tool_01KVWWW49FCJVE84Z9H6PH9N5M"
     )
 
     display_name: Optional[str] = None